Add geometric_sources.py
