+++ pyvista/core/utilities/geometric_sources.py
@@ -1,0 +1,252 @@
+"""Provides an easy way of generating several geometric sources.
+
+as well as some pure-python helpers.
+
+"""
+from typing import Sequence
+
+from pyvista.core import _vtk_core as _vtk
+from pyvista.core.utilities.misc import no_new_attr
+
+from .helpers import wrap
+
+
+@no_new_attr
+class ConeSource(_vtk.vtkConeSource):
+    """Cone source algorithm class.
+
+    Parameters
+    ----------
+    center : sequence[float], default: (0.0, 0.0, 0.0)
+        Center in ``[x, y, z]``. Axis of the cone passes through this
+        point.
+
+    direction : sequence[float], default: (1.0, 0.0, 0.0)
+        Direction vector in ``[x, y, z]``. Orientation vector of the
+        cone.
+
+    height : float, default: 1.0
+        Height along the cone in its specified direction.
+
+    radius : float, optional
+        Base radius of the cone.
+
+    capping : bool, optional
+        Enable or disable the capping the base of the cone with a
+        polygon.
+
+    angle : float, optional
+        The angle in degrees between the axis of the cone and a
+        generatrix.
+
+    resolution : int, default: 6
+        Number of facets used to represent the cone.
+
+    Examples
+    --------
+    Create a default ConeSource.
+
+    >>> import pyvista
+    >>> source = pyvista.ConeSource()
+    >>> source.output.plot(show_edges=True, line_width=5)
+    """
+
+    def __init__(
+        self,
+        center=(0.0, 0.0, 0.0),
+        direction=(1.0, 0.0, 0.0),
+        height=1.0,
+        radius=None,
+        capping=True,
+        angle=None,
+        resolution=6,
+    ):
+        """Initialize the cone source class."""
+        super().__init__()
+        self.center = center
+        self.direction = direction
+        self.height = height
+        self.capping = capping
+        if angle is not None and radius is not None:
+            raise ValueError("Both radius and angle specified. They are mutually exclusive.")
+        elif angle is not None and radius is None:
+            self.angle = angle
+        elif angle is None and radius is not None:
+            self.radius = radius
+        elif angle is not None and radius is None:
+            self.radius = 0.5
+        self.resolution = resolution
+
+    @property
+    def center(self) -> Sequence[float]:
+        """Get the center in ``[x, y, z]``. Axis of the cone passes through this point.
+
+        Returns
+        -------
+        sequence[float]
+            Center in ``[x, y, z]``. Axis of the cone passes through this
+            point.
+        """
+        return self.GetCenter()
+
+    @center.setter
+    def center(self, center: Sequence[float]):
+        """Set the center in ``[x, y, z]``. Axis of the cone passes through this point.
+
+        Parameters
+        ----------
+        center : sequence[float]
+            Center in ``[x, y, z]``. Axis of the cone passes through this
+            point.
+        """
+        self.SetCenter(center)
+
+    @property
+    def direction(self) -> Sequence[float]:
+        """Get the direction vector in ``[x, y, z]``. Orientation vector of the cone.
+
+        Returns
+        -------
+        sequence[float]
+            Direction vector in ``[x, y, z]``. Orientation vector of the
+            cone.
+        """
+        return self.GetDirection()
+
+    @direction.setter
+    def direction(self, direction: Sequence[float]):
+        """Set the center in ``[x, y, z]``. Axis of the cone passes through this point.
+
+        Parameters
+        ----------
+        direction : sequence[float]
+            Direction vector in ``[x, y, z]``. Orientation vector of the
+            cone.
+        """
+        self.SetDirection(direction)
+
+    @property
+    def height(self) -> float:
+        """Get the height along the cone in its specified direction.
+
+        Returns
+        -------
+        float
+            Height along the cone in its specified direction.
+        """
+        return self.GetHeight()
+
+    @height.setter
+    def height(self, height: float):
+        """Set the height of the cylinder.
+
+        Parameters
+        ----------
+        height : float
+            Height of the cylinder.
+        """
+        self.SetHeight(height)
+
+    @property
+    def radius(self) -> bool:
+        """Get base radius of the cone.
+
+        Returns
+        -------
+        float
+            Base radius of the cone.
+        """
+        return self.GetRadius()
+
+    @radius.setter
+    def radius(self, radius: float):
+        """Set base radius of the cone.
+
+        Parameters
+        ----------
+        radius : float
+            Base radius of the cone.
+        """
+        self.SetRadius(radius)
+
+    @property
+    def capping(self) -> bool:
+        """Enable or disable the capping the base of the cone with a polygon.
+
+        Returns
+        -------
+        bool
+            Enable or disable the capping the base of the cone with a
+            polygon.
+        """
+        return self.GetCapping()
+
+    @capping.setter
+    def capping(self, capping: bool):
+        """Set base capping of the cone.
+
+        Parameters
+        ----------
+        capping : bool, optional
+            Enable or disable the capping the base of the cone with a
+            polygon.
+        """
+        self.SetCapping(capping)
+
+    @property
+    def angle(self) -> float:
+        """Get the angle in degrees between the axis of the cone and a generatrix.
+
+        Returns
+        -------
+        float
+            The angle in degrees between the axis of the cone and a
+            generatrix.
+        """
+        return self.GetAngle()
+
+    @angle.setter
+    def angle(self, angle: float):
+        """Set the angle in degrees between the axis of the cone and a generatrix.
+
+        Parameters
+        ----------
+        angle : float, optional
+            The angle in degrees between the axis of the cone and a
+            generatrix.
+        """
+        self.SetAngle(angle)
+
+    @property
+    def resolution(self) -> int:
+        """Get number of points on the circular face of the cylinder.
+
+        Returns
+        -------
+        int
+            Number of points on the circular face of the cylinder.
+        """
+        return self.GetResolution()
+
+    @resolution.setter
+    def resolution(self, resolution: int):
+        """Set number of points on the circular face of the cylinder.
+
+        Parameters
+        ----------
+        resolution : int
+            Number of points on the circular face of the cylinder.
+        """
+        self.SetResolution(resolution)
+
+    @property
+    def output(self):
+        """Get the output data object for a port on this algorithm.
+
+        Returns
+        -------
+        pyvista.PolyData
+            Cone surface.
+        """
+        self.Update()
+        return wrap(self.GetOutput())
